Created useResumeData hook
Resume data state management
Profile data state management
Section ordering
Export settings
Loading, saving, and deleting resume data
All complex data transformation logic

Refactored ResumeEditor accordingly
